CFY-5317 Rollback idempotency test.
Test manager rollback: rerunning after a failed rollback, rolling back
twice.

Also allow using a pre-bootstrapped manager in the upgrade tests:

For debugging purposes, when working on the tests themselves, we can
use a pre-bootstrapped manager, to save time - so that we don't need
to bootstrap the manager every time as part of the test.